Added config framework file loading. Config file is in $XDG_CONFIG_HOME/archey.py.
The config file is non compulsary, and should be treated as dangerous.

Currently allowed config options are 'results_align' which can be 'top', 'bottom' or (default) 'center'. 'custom' is also allowed, but the config file must defign a 'custom_align' function taking the list of results, and a max_length kwarg, and returning the aligned results to be outputted.